+++ packages/flutter_deck/lib/src/controls/autoplay/autoplay.dart
@@ -1,1 +1,0 @@
-export 'flutter_deck_autoplay_notifier.dart';