+++ packages/flutter_deck/lib/src/controls/controls.dart
@@ -1,4 +1,3 @@
-export 'autoplay/autoplay.dart';
 export 'flutter_deck_controls.dart';
 export 'flutter_deck_controls_configuration.dart';
 export 'flutter_deck_controls_listener.dart';