+++ packages/flutter_deck/lib/src/plugins/autoplay/autoplay.dart
@@ -1,0 +1,3 @@
+export 'flutter_deck_autoplay_notifier.dart';
+export 'flutter_deck_autoplay_plugin.dart';
+export 'flutter_deck_autoplay_provider.dart';