+++ packages/flutter_deck/lib/src/plugins/plugins.dart
@@ -1,0 +1,2 @@
+export 'autoplay/autoplay.dart';
+export 'flutter_deck_plugin.dart';